spouts: Use trait for iterating items
No need to repeat the implementation everywhere
