+++ src/spouts/github/commits.php
@@ -13,6 +13,8 @@ use helpers\WebClient;
  * @author     Tim Gerundt <tim@gerundt.de>
  */
 class commits extends \spouts\spout {
+    use \helpers\ItemsIterator;
+
     /** @var string name of source */
     public $name = 'GitHub: commits';
 
@@ -44,81 +46,11 @@ class commits extends \spouts\spout {
         ]
     ];
 
-    /** @var ?array current fetched items */
-    protected $items = null;
-
     /** @var string global html url for the source */
     protected $htmlUrl = '';
 
     /** @var string URL of the favicon */
     protected $faviconUrl = 'https://assets-cdn.github.com/favicon.ico';
-
-    //
-    // Iterator Interface
-    //
-
-    /**
-     * reset iterator
-     *
-     * @return void
-     */
-    public function rewind() {
-        if ($this->items !== null) {
-            reset($this->items);
-        }
-    }
-
-    /**
-     * receive current item
-     *
-     * @return \SimplePie_Item current item
-     */
-    public function current() {
-        if ($this->items !== null) {
-            return $this;
-        }
-
-        return false;
-    }
-
-    /**
-     * receive key of current item
-     *
-     * @return mixed key of current item
-     */
-    public function key() {
-        if ($this->items !== null) {
-            return key($this->items);
-        }
-
-        return false;
-    }
-
-    /**
-     * select next item
-     *
-     * @return \SimplePie_Item next item
-     */
-    public function next() {
-        if ($this->items !== null) {
-            next($this->items);
-        }
-
-        return $this;
-    }
-
-    /**
-     * end reached
-     *
-     * @return bool false if end reached
-     */
-    public function valid() {
-        if ($this->items !== null) {
-            return current($this->items) !== false;
-        }
-
-        return false;
-    }
 
     //
     // Source Methods
