+++ src/spouts/reddit/reddit2.php
@@ -17,6 +17,8 @@ use Stringy\Stringy as S;
  * @author     Tobias Zeising <tobias.zeising@aditu.de>
  */
 class reddit2 extends \spouts\spout {
+    use \helpers\ItemsIterator;
+
     /** @var string name of spout */
     public $name = 'Reddit';
 
@@ -50,9 +52,6 @@ class reddit2 extends \spouts\spout {
 
     /** @var string the reddit_session cookie */
     private $reddit_session = '';
-
-    /** @var array|null current fetched items */
-    protected $items = null;
 
     /** @var string favicon url */
     private $faviconUrl = '';
@@ -92,73 +91,6 @@ class reddit2 extends \spouts\spout {
         }
 
         $this->items = $json['data']['children'];
-    }
-
-    //
-    // Iterator Interface
-    //
-
-    /**
-     * reset iterator
-     *
-     * @return void
-     */
-    public function rewind() {
-        if ($this->items !== null) {
-            reset($this->items);
-        }
-    }
-
-    /**
-     * receive current item
-     *
-     * @return \SimplePie_Item current item
-     */
-    public function current() {
-        if ($this->items !== null) {
-            return $this;
-        }
-
-        return false;
-    }
-
-    /**
-     * receive key of current item
-     *
-     * @return mixed key of current item
-     */
-    public function key() {
-        if ($this->items !== null) {
-            return key($this->items);
-        }
-
-        return null;
-    }
-
-    /**
-     * select next item
-     *
-     * @return \SimplePie_Item next item
-     */
-    public function next() {
-        if ($this->items !== null) {
-            next($this->items);
-        }
-
-        return $this;
-    }
-
-    /**
-     * end reached
-     *
-     * @return bool false if end reached
-     */
-    public function valid() {
-        if ($this->items !== null) {
-            return current($this->items) !== false;
-        }
-
-        return false;
     }
 
     /**
